fix : file relocate
