+++ moa2/src/main/java/com/moa2/controller/auth/AuthController.java
@@ -3,7 +3,7 @@ package com.moa2.controller.auth;
 import com.moa2.domain.member.Member;
 import com.moa2.dto.auth.LoginDto;
 import com.moa2.dto.auth.SignupDto;
-import com.moa2.security.jwt.TokenDto;
+import com.moa2.dto.auth.TokenDto;
 import com.moa2.service.auth.AuthService;
 import com.moa2.service.member.MemberService;
 import jakarta.validation.Valid;
