+++ moa2/src/main/java/com/moa2/dto/auth/TokenDto.java
@@ -1,4 +1,4 @@
-package com.moa2.security.jwt;
+package com.moa2.dto.auth;
 
 import lombok.Getter;
