+++ moa2/src/main/java/com/moa2/security/userdetails/MemberDetailsService.java
@@ -1,4 +1,4 @@
-package com.moa2.service.userdetails;
+package com.moa2.security.userdetails;
 
 import com.moa2.domain.member.Member;
 import com.moa2.repository.member.MemberRepository;
